Pharmacy mobile app
This is a project to develop a mobile application ios for a pharmacy. Our application will allow customers to consult the list of all pharmacies available on the database and to be able to search for drugs and place an order. When the order is placed, the customer is given a qrcode that will allow him to go to the pharmacy to retrieve his order and proceed to payment. He will also be able to consult the pharmacy on call